sam0/uart: merge samd21 & saml21 uart driver
