+++ cpu/samd21/include/periph_cpu.h
@@ -89,18 +89,6 @@ typedef struct {
 } pwm_conf_t;
 
 /**
- * @brief   UART device configuration
- */
-typedef struct {
-    SercomUsart *dev;       /**< pointer to the used UART device */
-    gpio_t rx_pin;          /**< pin used for RX */
-    gpio_t tx_pin;          /**< pin used for TX */
-    gpio_mux_t mux;         /**< alternative function for pins */
-    uart_rxpad_t rx_pad;    /**< pad selection for RX line */
-    uart_txpad_t tx_pad;    /**< pad selection for TX line */
-} uart_conf_t;
-
-/**
  * @brief   Return the numeric id of a SERCOM device derived from its address
  *
  * @param[in] sercom    SERCOM device
